livraison_nourriture/requirements.txt
Django==5.1.4
cloudinary==1.41.0
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.2
psycopg2-binary==2.9.10
pymesomb==2.0.3
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
supabase==2.10.0
django-cloudinary-storage==0.3.0
django-crontab
django-q2
setuptools
mysqlclient